requirements.txt
discord.py>=1.6.0
requests>=2.27.1
ornhub_api>=0.2.0